# Use bitset in LAD and clean up code