--- a/tests/CMakeLists.txt
+++ b/tests/CMakeLists.txt
@@ -1018,6 +1018,7 @@ add_benchmarks(
   igraph_vertex_connectivity
   igraph_voronoi
   inc_vs_adj
+  lad
 )
 
 # dot product