… tests/benchmarks/lad.c