Benchmark before:

```
| 1 Kautz(3,3) 10x                                                                 0.603s   0.57s  0.028s
| 2 Kautz(3,4) 3x                                                                   0.85s  0.829s  0.015s
| 3 Kautz(4,3) 1x                                                                   1.03s      1s  0.022s
```

After:

```
| 1 Kautz(3,3) 10x                                                                 0.597s  0.567s  0.028s
| 2 Kautz(3,4) 3x                                                                  0.807s   0.79s  0.014s
| 3 Kautz(4,3) 1x                                                                  0.942s  0.916s  0.022s
```