@szhorvat 2024-05-25
@ntamas Please have a careful look at this commit: https://github.com/igraph/igraph/pull/2617/commits/72ed68c1cd0c0db86eeaad6f90b7c09a66819f83 I think the memory management of `mv` and `mu` was flawed. Why did it do an `igraph_free(val);` manually when `val` was already on the finally stack?