@ntamas 2024-05-27
Yes, the memory management was implemented incorrecrtly there, thanks for fixing it!